requirements.txt
alembic==1.0.10
Click==7.0
DBUtils==1.3
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.5.1
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SocketIO==4.1.0
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
python-engineio==3.8.1
python-socketio==4.0.2
redis==3.2.1
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.4
